Move `JSONResponse` class
