+++ formwork/src/Admin/Admin.php
@@ -8,13 +8,13 @@ use Formwork\Admin\Users\Users;
 use Formwork\Assets;
 use Formwork\Formwork;
 use Formwork\Page;
+use Formwork\Response\JSONResponse;
 use Formwork\Router\RouteParams;
 use Formwork\Router\Router;
 use Formwork\Translations\Translation;
 use Formwork\Utils\FileSystem;
 use Formwork\Utils\Header;
 use Formwork\Utils\HTTPRequest;
-use Formwork\Utils\JSONResponse;
 use Formwork\Utils\Notification;
 use Formwork\Utils\Session;
 use Formwork\Utils\Str;
