+++ formwork/src/Admin/Controllers/PagesController.php
@@ -11,11 +11,11 @@ use Formwork\Formwork;
 use Formwork\Languages\LanguageCodes;
 use Formwork\Page;
 use Formwork\Parsers\YAML;
+use Formwork\Response\JSONResponse;
 use Formwork\Router\RouteParams;
 use Formwork\Site;
 use Formwork\Utils\FileSystem;
 use Formwork\Utils\HTTPRequest;
-use Formwork\Utils\JSONResponse;
 use Formwork\Utils\Session;
 use Formwork\Utils\Str;
 use Formwork\Utils\Uri;
